Initial commit:
This is a library project to calculate the suns movement
I use it to optimize solar array orientation, and thus, power production.

main.c
Contains testcases

soltrack.c
Contains utility to calculate the suns movement
Right now, only the initialisation of the axial tilt matrix and the unit
vector orthogonal to the earths surface are done.

soltrack.h
The public utilities that can be used

What is to be done:
Define versioning
Include license into every file
Big part of the math
Finally, make a nice explanation why it should work